feat: refactor getPageContent - extract getPageByIdAndConfigNavbarAndFooter to a file and rename it to getPageById   - move configNavbar to the getPageById folder   - move configFooter to the getPageById folder   - move insertSection to the getPageById folder
