+++ packages/nextjs-contentful-renderer/src/getPageContent/getPageContent.ts
@@ -1,8 +1,6 @@
-import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
 import { GetServerSidePropsContext } from 'next';
-import { configNavbar } from '../configNavbar';
-import { configFooter } from '../configFooter';
-import { getPageById, getRouteBySlug } from '../utils/contentful';
+import { getPageById } from '../utils/getPageById';
+import { getRouteBySlug } from '../utils/contentful';
 import { Page } from '../utils/contentful/getPageById/types';
 import { getVisitedPageIdFromCookies, setCookie } from '../utils/cookies';
 import { getApolloClient } from '../utils/getApolloClient';
@@ -21,7 +19,7 @@ export const getPageContent = async (
 
   const existingPageId = getVisitedPageIdFromCookies(context, slug);
   if (existingPageId) {
-    const page = await getPageByIdAndConfigNavbarAndFooter(apolloClient, existingPageId, preview);
+    const page = await getPageById(apolloClient, existingPageId, preview);
     if (page) {
       return page;
     }
@@ -31,21 +29,8 @@ export const getPageContent = async (
   if (!route) return undefined;
 
   const pageId = getPageId(route, context.query.utm_campaign);
-  const page = await getPageByIdAndConfigNavbarAndFooter(apolloClient, pageId, preview);
+  const page = await getPageById(apolloClient, pageId, preview);
   setCookie(context, slug, pageId);
 
-  return page;
-};
-
-const getPageByIdAndConfigNavbarAndFooter = async (
-  apolloClient: ApolloClient<NormalizedCacheObject>,
-  pageId: string,
-  preview: boolean
-) => {
-  const page = await getPageById(apolloClient, pageId, preview);
-  if (page) {
-    page.content = configNavbar(page);
-    page.content = configFooter(page);
-  }
   return page;
 };
