+++ packages/nextjs-contentful-renderer/src/utils/getPageById/configFooter/configFooter.test.ts
