+++ packages/nextjs-contentful-renderer/src/utils/getPageById/configFooter/configFooter.ts
@@ -1,0 +1,4 @@
+import { insertSection } from '../insertSection';
+import { Block, Page } from '../../contentful/getPageById/types';
+
+export const configFooter = (page: Partial<Page>): Block[] => insertSection(page.content, page.footer, 'push');
