+++ packages/nextjs-contentful-renderer/src/utils/getPageById/configFooter/index.ts
