+++ packages/nextjs-contentful-renderer/src/utils/getPageById/configNavbar/configNavbar.ts
@@ -1,0 +1,4 @@
+import { Block, Page } from '../../contentful/getPageById/types';
+import { insertSection } from '../insertSection';
+
+export const configNavbar = (page: Partial<Page>): Block[] => insertSection(page.content, page.navbar, 'unshift');
