+++ packages/nextjs-contentful-renderer/src/utils/getPageById/index.ts
@@ -1,0 +1,1 @@
+export * from './getPageById';
